more tests + more pythonic winml tutorial
